This is a fancy textview for iOS
This project is base on EGOTextView, and add some feature to support
text and image emotion editing at the same time.

In BBS, we offten use [a emotion] to add a image, say emotion001.png.
Here you only have to set up a text-image mapping, it's very easy to
use.